Rebase from Master

+++ pkg/storage/objects/store.go
@@ -8,16 +8,20 @@ import (
 	"io"
 	"time"
 
-	"github.com/gogo/protobuf/proto"
+	"github.com/golang/protobuf/proto"
+	"github.com/zeebo/errs"
 	"go.uber.org/zap"
 	monkit "gopkg.in/spacemonkeygo/monkit.v2"
 
-	"storj.io/storj/pkg/paths"
 	"storj.io/storj/pkg/ranger"
 	"storj.io/storj/pkg/storage/streams"
+	"storj.io/storj/pkg/storj"
 )
 
 var mon = monkit.Package()
+
+// NoPathError is an error class for missing object path
+var NoPathError = errs.Class("no object path specified")
 
 // Meta is the full object metadata
 type Meta struct {
@@ -30,22 +34,18 @@ type Meta struct {
 
 // ListItem is a single item in a listing
 type ListItem struct {
-	Path     paths.Path
+	Path     storj.Path
 	Meta     Meta
 	IsPrefix bool
 }
 
 // Store for objects
 type Store interface {
-	Meta(ctx context.Context, path paths.Path) (meta Meta, err error)
-	Get(ctx context.Context, path paths.Path) (rr ranger.RangeCloser,
-		meta Meta, err error)
-	Put(ctx context.Context, path paths.Path, data io.Reader,
-		metadata SerializableMeta, expiration time.Time) (meta Meta, err error)
-	Delete(ctx context.Context, path paths.Path) (err error)
-	List(ctx context.Context, prefix, startAfter, endBefore paths.Path,
-		recursive bool, limit int, metaFlags uint32) (items []ListItem,
-		more bool, err error)
+	Meta(ctx context.Context, path storj.Path) (meta Meta, err error)
+	Get(ctx context.Context, path storj.Path) (rr ranger.Ranger, meta Meta, err error)
+	Put(ctx context.Context, path storj.Path, data io.Reader, metadata SerializableMeta, expiration time.Time) (meta Meta, err error)
+	Delete(ctx context.Context, path storj.Path) (err error)
+	List(ctx context.Context, prefix, startAfter, endBefore storj.Path, recursive bool, limit int, metaFlags uint32) (items []ListItem, more bool, err error)
 }
 
 type objStore struct {
@@ -57,28 +57,39 @@ func NewStore(store streams.Store) Store {
 	return &objStore{s: store}
 }
 
-func (o *objStore) Meta(ctx context.Context, path paths.Path) (meta Meta,
-	err error) {
+func (o *objStore) Meta(ctx context.Context, path storj.Path) (meta Meta, err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if len(path) == 0 {
+		return Meta{}, NoPathError.New("")
+	}
+
 	m, err := o.s.Meta(ctx, path)
 	return convertMeta(m), err
 }
 
-func (o *objStore) Get(ctx context.Context, path paths.Path) (
-	rr ranger.RangeCloser, meta Meta, err error) {
+func (o *objStore) Get(ctx context.Context, path storj.Path) (
+	rr ranger.Ranger, meta Meta, err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if len(path) == 0 {
+		return nil, Meta{}, NoPathError.New("")
+	}
+
 	rr, m, err := o.s.Get(ctx, path)
 	return rr, convertMeta(m), err
 }
 
-func (o *objStore) Put(ctx context.Context, path paths.Path, data io.Reader,
-	metadata SerializableMeta, expiration time.Time) (meta Meta, err error) {
+func (o *objStore) Put(ctx context.Context, path storj.Path, data io.Reader, metadata SerializableMeta, expiration time.Time) (meta Meta, err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if len(path) == 0 {
+		return Meta{}, NoPathError.New("")
+	}
 
 	// TODO(kaloyan): autodetect content type
 	// if metadata.GetContentType() == "" {}
 
-	// TODO(kaloyan): encrypt metadata.UserDefined before serializing
 	b, err := proto.Marshal(&metadata)
 	if err != nil {
 		return Meta{}, err
@@ -87,13 +98,17 @@ func (o *objStore) Put(ctx context.Context, path paths.Path, data io.Reader,
 	return convertMeta(m), err
 }
 
-func (o *objStore) Delete(ctx context.Context, path paths.Path) (err error) {
+func (o *objStore) Delete(ctx context.Context, path storj.Path) (err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if len(path) == 0 {
+		return NoPathError.New("")
+	}
+
 	return o.s.Delete(ctx, path)
 }
 
-func (o *objStore) List(ctx context.Context, prefix, startAfter,
-	endBefore paths.Path, recursive bool, limit int, metaFlags uint32) (
+func (o *objStore) List(ctx context.Context, prefix, startAfter, endBefore storj.Path, recursive bool, limit int, metaFlags uint32) (
 	items []ListItem, more bool, err error) {
 	defer mon.Task()(&ctx)(&err)
